+++ HISTORY.rst
@@ -21,6 +21,7 @@
 * Open staging file in explicit binary mode. (Python 3)
 * Add separate Python 2 and Python 3 versions of the 
   MetadataType model .comma_splitter() static method.
+* Update the metadata app tests to work on Python 3.
 
 3.1.4 (2018-10-4)
 =================